+++ src/SidebarTop.tsx
@@ -34,7 +34,7 @@ export const SidebarTop = ({ api }: SidebarTopProps) => {
   const [localBuildProgress] = useSharedState<LocalBuildProgress>(LOCAL_BUILD_PROGRESS);
   const isRunning =
     !!localBuildProgress &&
-    !["aborted", "complete", "error"].includes(localBuildProgress.currentStep);
+    !["aborted", "complete", "error", "limited"].includes(localBuildProgress.currentStep);
 
   const [configInfo] = useSharedState<ConfigInfoPayload>(CONFIG_INFO);
   const hasConfigProblem = Object.keys(configInfo?.problems || {}).length > 0;
@@ -125,6 +125,23 @@ export const SidebarTop = ({ api }: SidebarTopProps) => {
         link: undefined,
       });
     }
+
+    if (localBuildProgress?.currentStep === "limited") {
+      addNotification({
+        id: `${ADDON_ID}/build-limited`,
+        content: {
+          headline: "Build limited",
+          subHeadline:
+            "Your account has insufficient snapshots remaining to run this build. Visit your billing page to find out more.",
+        },
+        icon: {
+          name: "failed",
+          color: color.negative,
+        },
+        // @ts-expect-error SB needs a proper API for no link
+        link: undefined,
+      });
+    }
   }, [
     addNotification,
     clearNotification,
@@ -143,6 +160,10 @@ export const SidebarTop = ({ api }: SidebarTopProps) => {
   if (!isLoggedIn) warning = "Visual tests locked until you are logged in.";
   if (gitInfoError) warning = "Visual tests locked due to Git synchronization problem.";
   if (hasConfigProblem) warning = "Visual tests locked due to configuration problem.";
+
+  if (global.CONFIG_TYPE !== "DEVELOPMENT") {
+    return null;
+  }
 
   return (
     <SidebarTopButton