+++ src/screens/Authentication/SetSubdomain.tsx
@@ -57,7 +57,7 @@ export const SetSubdomain = ({ onBack, onSignIn }: SetSubdomainProps) => {
   );
 
   return (
-    <Screen footer={null}>
+    <Screen footer={null} ignoreConfig>
       <Container>
         <BackButton onClick={onBack}>
           <ChevronLeftIcon color={theme.base === "light" ? "currentColor" : theme.color.medium} />