+++ src/screens/Errors/GitNotFound.tsx
@@ -1,41 +1,15 @@
-import { styled } from "@storybook/theming";
 import React from "react";
 
+import { Box, BoxContent, BoxTitle } from "../../components/Box";
 import { Code } from "../../components/Code";
 import { Container } from "../../components/Container";
 import { Icon, Link } from "../../components/design-system";
 import { Heading } from "../../components/Heading";
 import { VisualTestsIcon } from "../../components/icons/VisualTestsIcon";
-import { Section } from "../../components/layout";
 import { Screen } from "../../components/Screen";
 import { Stack } from "../../components/Stack";
 import { Text } from "../../components/Text";
 import { useUninstallAddon } from "../Uninstalled/UninstallContext";
-
-const InfoSection = styled(Section)(({ theme }) => ({
-  display: "flex",
-  flexDirection: "row",
-  alignItems: "center",
-  borderRadius: theme.appBorderRadius,
-  background: theme.base === "light" ? theme.color.lightest : theme.color.darkest,
-  border: `1px solid ${theme.appBorderColor}`,
-  padding: 15,
-  flex: 1,
-
-  maxWidth: "500px",
-}));
-
-const InfoSectionText = styled(Text)(({ theme }) => ({
-  marginLeft: 14,
-  flex: 1,
-  textAlign: "left",
-  color: theme.base === "light" ? theme.color.darker : theme.color.lighter,
-}));
-
-const InfoSectionTextTitle = styled.b(() => ({
-  display: "block",
-  marginBottom: 2,
-}));
 
 export const GitNotFound = () => {
   const { uninstallAddon } = useUninstallAddon();
@@ -51,15 +25,15 @@ export const GitNotFound = () => {
               changes.
             </Text>
           </div>
-          <InfoSection>
+          <Box>
             <Icon icon="lock" />
-            <InfoSectionText>
-              <InfoSectionTextTitle>Git not detected</InfoSectionTextTitle>
+            <BoxContent>
+              <BoxTitle>Git not detected</BoxTitle>
               This addon requires Git to associate test results with commits and branches. Run{" "}
               <Code>git init</Code> and make your first commit
               <Code>git commit -m</Code> to get started!
-            </InfoSectionText>
-          </InfoSection>
+            </BoxContent>
+          </Box>
           <Link
             target="_blank"
             href="https://www.chromatic.com/docs/visual-tests-addon#git-addon"