+++ src/screens/LinkProject/LinkProject.stories.ts
@@ -182,7 +182,7 @@ const manyProjects = {
         ],
       },
       {
-        id: "account:4567",
+        id: "account:4568",
         name: "below the fold corp",
         newProjectUrl: "https://www.chromatic.com/apps?accountId=4567&view=createProject",
         projects: [