+++ src/screens/VisualTests/SnapshotControls.stories.tsx
@@ -30,7 +30,6 @@ const meta = {
   ],
   args: {
     isOutdated: false,
-    startDevBuild: action("startDevBuild"),
   },
   parameters: {
     chromatic: {
@@ -51,12 +50,6 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default = {} satisfies Story;
-
-export const Outdated = {
-  args: {
-    isOutdated: true,
-  },
-} satisfies Story;
 
 export const InProgress = {
   parameters: {